pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-deploy-stores"
version = "1.0.0"
description = "Auto-deploy pipeline for Apple App Store and Google Play Store"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MIDNGHTSAPPHIRE", email = "dev@glowstarlabs.com"}
]
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "cryptography>=42.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.3",
    "pillow>=10.2.0",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.6",
    "jsonschema>=4.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
autodeploy = "cli.main:cli"

[project.entry-points."autodeploy.plugins"]

[tool.setuptools.packages.find]
include = ["cli*", "mcp_server*", "vault*", "wrapper*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
